+++ src/audio.rs
@@ -42,36 +42,14 @@ pub struct Audio<S: Sample> {
 }
 
 impl<S: Sample> Audio<S> {
-    /// Get view of samples.
+    /// Get view of samples as a slice.
     pub fn as_slice(&self) -> &[S] {
         &self.samples
     }
     
-    /// Get view of samples.
+    /// Get view of samples as a mutable slice.
     pub fn as_slice_mut(&mut self) -> &mut [S] {
         &mut self.samples
-    }
-
-    /// Get view of samples as a `u8` slice.
-    #[allow(unsafe_code)]
-    pub fn as_u8_slice(&self) -> &[u8] {
-        unsafe {
-            let (prefix, v, suffix) = self.samples.align_to::<u8>();
-            debug_assert!(prefix.is_empty());
-            debug_assert!(suffix.is_empty());
-            v
-        }
-    }
-
-    /// Get view of samples as a mutable `u8` slice.
-    #[allow(unsafe_code)]
-    pub fn as_u8_slice_mut(&mut self) -> &mut [u8] {
-        unsafe {
-            let (prefix, v, suffix) = self.samples.align_to_mut::<u8>();
-            debug_assert!(prefix.is_empty());
-            debug_assert!(suffix.is_empty());
-            v
-        }
     }
 
     /// Returns an iterator over the samples.